[snmp/agent] Add default value for mib_storage for sub-agent
When starting a sub-agent we previously did not provide
a value for mib_storage, which was alright because ets
was assumed as a default in every place where it was used.
Now we expect the value to be defined and therefor we
must explicitly add the default value for sub-agents when
staring them.