[snmp/agent] Changed the test suite expect function
There was an id as the first agument to each expect call.
Ufortunatly, this id was often 1 and since the same test
function(s) was called in many test cases, there was many
"Expect 1", and therefor no way of knowing which expect
actually was performed.
The expect functions has been changed to instead take module
and line number.